requirements.txt
Django==3.2.7
django-environ==0.7.0
gunicorn==20.0.4
psycopg2==2.9.1
whitenoise==5.3.0
asgiref==3.4.1
backcall==0.2.0
colorama==0.4.4
cycler==0.10.0
decorator==5.1.0
dj-database-url==0.5.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
ipython==7.28.0
jedi==0.18.0
joblib==1.1.0
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.21.2
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.20
Pygments==2.10.0
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
traitlets==5.1.0
wcwidth==0.2.5

